Remodeled object and component classes somewhat. Component vector moved to Component class and Object class now has a root_component member variable.
